Initial commit
rootDir is set within the .apytly.conf file to ensure the right
directory is used when serving the apt repository locally. As otherwise,
aptly will only serve repositories found in the ~/.aptly/public
directory. The rootDir will also contain a public directory with a
symbolic link inside it that points to the top-level deb directory, so
the URL used in sources.list(5) files can remain short. Jekyll is
configured via the _config.yml file to ignore this symbolic link when
building because otherwise it will replace this with the directory and
its contents. Inflating the size of the artifacts by 2x.

When downloading the kubectl packages, I anticipate that the apt-get
executable will install the most recent package (that is, the one with
the higher version number) for each version. This is even though all
the kubectl packages will have the same priority assigned to them by
apt.

Even though kubectl's official installation documentation recommends
using '/' as the suite in their sources.list(5) file, I've chosen to set
the suite (or aptly's distribution option) for the apt repository as
'all'. I believe both to convey the same meaning, that is, that the
repository's packages are suite generic, but the latter is clearer on
this. This suite is also used by the apt repository recommended by
Helm's documentation.

There is a recommended apt executable version to shorten the number of
packages required to install the kubectl packages when following the
official installation documentation. After apt version 1.5, HTTPS has
been integrated into the apt package, voiding the need for the apt
transport https package. Also, support for embedded GPG/PGP keys were
added as of version 2.3.10, voiding the need for the gnupg package.